Classe Ler_DataBase para ler o ficheiro quase concluida(falta rever) e alterados alguns atributos na classe Terreno

+++ ES-2024-1Sem-LETI-GrupoH/src/main/java/Terreno.java
@@ -1,14 +1,14 @@
 public class Terreno {
 
     private int objectID;
-    private int parID;
-    private int parNUM;
-    private int shape_Length;
-    private int shape_Area;
+    private double parID;
+    private long parNUM;
+    private double shape_Length;
+    private double shape_Area;
     private String geometry;
     private int owner;
 
-    public Terreno(int objectID, int parID, int parNUM, int shape_Length, int shape_Area, String geometry, int owner) {
+    public Terreno(int objectID, double parID, long parNUM, double shape_Length, double shape_Area, String geometry, int owner) {
         this.objectID = objectID;
         this.parID = parID;
         this.parNUM = parNUM;
